requirements.txt
appdirs==1.4.3
Babel==2.3.4
Django==1.11
django-appconf==1.0.2
django-compressor==2.1.1
django-extra-views==0.6.4
django-haystack==2.5.1
django-oscar==1.4
django-tables2==1.4.2
django-treebeard==4.1.0
django-widget-tweaks==1.4.1
factory-boy==2.8.1
Faker==0.7.9
funcsigs==1.0.2
ipaddress==1.0.18
mock==2.0.0
olefile==0.44
packaging==16.8
pbr==2.0.0
phonenumbers==8.3.2
Pillow==4.0.0
purl==1.3
pycountry==17.1.8
pyparsing==2.2.0
pysolr==3.6.0
python-dateutil==2.6.0
pytz==2017.2
rcssmin==1.0.6
requests==2.13.0
rjsmin==1.0.12
six==1.10.0
sorl-thumbnail==12.4a1
Unidecode==0.4.20
whitenoise==3.3.0
